+++ example/audio-identify.js
@@ -1,4 +1,4 @@
-const audio = require('../functions/services/audio')
+const audio = require('../services/audio')
 
 audio.identify('./example20mb.wav').then(result => {
   console.log(JSON.stringify(result))