+++ routes/metrics.js
@@ -1,0 +1,12 @@
+const { PROMETHEUS_ENABLED, register } = require('../utils/prometheus')
+if (PROMETHEUS_ENABLED) {
+  const router = require('express').Router()
+
+  router.route('/').get(async (req, res) => {
+    res.setHeader('Content-Type', register.contentType)
+    res.send(await register.metrics())
+    register.resetMetrics()
+  })
+
+  module.exports = router
+}